requirements.txt
anyio==3.7.1
asttokens 
backcall 
blinker==1.8.2
boto3==1.24.84
botocore==1.27.96
botocore-stubs==1.27.86
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
colorama 
comm 
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
debugpy
decorator 
dill==0.3.5.1
dnspython==2.2.1
evidently==0.1.58.dev0
exceptiongroup==1.3.0
executing 
fastapi==0.78.0
Flask==3.0.3
fonttools==4.57.0
from-root==1.0.2
h11==0.16.0
httptools==0.5.0
idna==3.10
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata 
importlib_resources==6.4.5
ipykernel 
ipython 
itsdangerous==2.2.0
jedi 
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jupyter_client 
jupyter_core 
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline 
mypy-boto3-s3==1.24.76
narwhals==1.42.1
nest_asyncio 
neuro-mf==0.0.5
numpy==1.24.4
packaging
pandas==2.0.3
parso 
patsy==1.0.1
pickleshare 
pillow==10.4.0
pip-chill==1.0.1
platformdirs 
plotly==6.2.0
prompt_toolkit 
psutil 
pure_eval
pydantic==1.10.22
Pygments 
pymongo==4.2.0
pyparsing==3.1.4
python-dateutil 
python-dotenv==0.21.0
python-multipart==0.0.20
pytz==2025.2
pywin32==306
PyYAML==5.4.1
pyzmq 
requests==2.32.4
s3transfer==0.6.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six 
sniffio==1.3.1
stack-data 
starlette==0.19.1
statsmodels==0.14.1
threadpoolctl==3.5.0
tornado 
traitlets 
types-awscrt==0.27.5
types-s3transfer==0.6.0.post4
typing_extensions 
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.18.3
watchfiles==0.17.0
wcwidth 
websockets==10.3
Werkzeug==3.0.6
wincertstore==0.2
xgboost==1.6.2
zipp 
